\name{PointPatternAnalysis}
\alias{PointPatternAnalysis}
\title{
Basic spatial analysis using Point pattern analysis
}
\description{
Offers instruction on how to load two types of spatial data, vectors and rasters, and plot them.  Puts data coercion into useful formats for spatial analysis.  Explores spatial dependency based on a) between-point interaction, b) external covariates.
}
\usage{
PointPatternAnalysis(x)
}
\value{
  \item{main.sites }{Loading and plotting spatial data}
  \item{plot }{graph of elevation}
  \item{plot }{graph of main site}
  \item{plot }{graph of secondary site}
  \item{legend }{identifier of details in graphs}
}
\references{
Riris, P. (n.d.). Basic spatial analysis in R: Point pattern analysis.  Retrieved from https://benmarwick.github.io/How-To-Do-Archaeological-Science-Using-R/basic-spatial-analysis-in-r-point-pattern-analysis.html
}
\author{
Elizabeth Siniscalchi
\note{
These functions are intended to be generic so that they may be used by anyone rather than with a specific set of archaeological data.
}
\keyword{ Spatial }
